// src/types.ts
export type Options = {
    databaseUrl?: string;
    authentication?: string;
    // requestQuery?: Query;
    tokenAuthentication?: boolean,
}

/* export enum Query {
    ID_AUTH = '.json?auth=',
    TOKEN_AUTH = '.json?access_token=',
    SHALLOW_QUERY = '.json?shallow=true'
}
FOR ANOTHER DAY
*/